added a codewar kata and ints
There was things about ints in a tutorial I was doing. It was a bit
tricky, but I really like the idea of learning c so I will keep
practising. I tried a really easy kata but I couldn't understand it in
c. I also would like to have a makefile for this at one point too, but
that is a bit challenging also. I am happy with my work.