Added functions (prototypes) and loops
I am slowly starting to appreciate c. It is a language that does have a
lot of syntax to learn compared to python, but that makes it a lot
easier to do things beyond the very basics. Everything that is done
makes a lot of sense.

I learnt functions and loops.